+++ mobile/App.tsx
@@ -3,4 +3,5 @@ import 'react-native-gesture-handler';
 import { StatusBar } from 'expo-status-bar';
 import { AuthProvider } from './src/auth/AuthProvider';
 import RootNavigator from './src/navigation/RootNavigator';
-export default function App(){return <AuthProvider><StatusBar style="light"/><RootNavigator/></AuthProvider>}
+import { LanguageButton,LanguageProvider } from './src/i18n/MobileLanguage';
+export default function App(){return <LanguageProvider><AuthProvider><StatusBar style="light"/><RootNavigator/><LanguageButton/></AuthProvider></LanguageProvider>}